// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      properties: {
        Row: {
          address: string
          amenities: string[] | null
          bathrooms: number | null
          bedrooms: number | null
          city: string
          country: string
          created_at: string
          currency: string | null
          description: string | null
          host_id: string
          id: string
          images: string[] | null
          is_available: boolean | null
          latitude: number | null
          longitude: number | null
          max_guests: number | null
          price_per_month: number
          property_type: string
          title: string
          updated_at: string
        }
        Insert: {
          address: string
          amenities?: string[] | null
          bathrooms?: number | null
          bedrooms?: number | null
          city: string
          country: string
          created_at?: string
          currency?: string | null
          description?: string | null
          host_id: string
          id?: string
          images?: string[] | null
          is_available?: boolean | null
          latitude?: number | null
          longitude?: number | null
          max_guests?: number | null
          price_per_month: number
          property_type: string
          title: string
          updated_at?: string
        }
        Update: {
          address?: string
          amenities?: string[] | null
          bathrooms?: number | null
          bedrooms?: number | null
          city?: string
          country?: string
          created_at?: string
          currency?: string | null
          description?: string | null
          host_id?: string
          id?: string
          images?: string[] | null
          is_available?: boolean | null
          latitude?: number | null
          longitude?: number | null
          max_guests?: number | null
          price_per_month?: number
          property_type?: string
          title?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "properties_host_id_fkey"
            columns: ["host_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["user_id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}